# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "文件统一上传API文档",
        "contact": {
            "name": "Gong Zhigang",
            "email": "zggong@aibee.com"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/upload": {
            "post": {
                "description": "上传文件接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否为多文件上传",
                        "name": "isMultipart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "上传文件类型 images/file/media",
                        "name": "upload_type",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "文件集",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/web3/ipfs": {
            "get": {
                "description": "查询文件接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件集",
                        "name": "cid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "上传文件接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否为多文件上传",
                        "name": "isMultipart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "上传文件类型 images/file/media",
                        "name": "upload_type",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "文件集",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        }
    }
}